Update imports with dataclass refactor
