+++ tests/test_api/test_ai/test_priority_algorithm/test_mutations/test_greedy_random_local_max.py
@@ -2,8 +2,8 @@ import random
 import unittest
 from collections import Counter
 
-from api.ai.priority_algorithm.custom_models import PriorityTeam, PriorityTeamSet
-from api.models.team import TeamShell
+from api.ai.priority_algorithm.custom_dataclasses import PriorityTeam, PriorityTeamSet
+from api.dataclasses.team import TeamShell
 from benchmarking.data.simulated_data.mock_student_provider import (
     MockStudentProvider,
     MockStudentProviderSettings,
